Fixed myjobentry
Oh man, that was painful. tabular rows do not like it when you
wrap {} around everything with \\ for a newline within the brackets.

Removed the brackets wrapping the line and have the \\ on the outside,
now things work. I'll make it prettier in a future commit (or at least
make the statements better)